api/requirements.txt
fastapi[standard]
pydantic
typing
requests
urllib3
python-jose[cryptography]

google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

redis[hiredis]
psycopg[binary]

isort[requirements_deprecated_finder]